refactoring the call page to use view models and removing one of the sql calls
